requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-openai
langchain-text-splitters
faiss-cpu
sentence-transformers
datasets
pandas
python-dotenv
streamlit
ragas
openai
google-generativeai